requirement_pip.txt
# pip install -r requirement_pip.txt
django
djangorestframework
django-filter
django-cors-headers
djangorestframework-jwt
